Issue 11457: create new user roles for EHR
Let Roles decide if they apply to a given resource. Use this to filter the set of roles we show in the regular project/folder permissions UI.

Allow modules to define roles that apply to datasets, and show them as options if the study uses advanced security with editable datasets.

SVN r15661 |2011-02-25 22:15:01 +0000